게시물(post) 등록, 수정, 삭제, 조회, 특정 게시물 조회 구현

+++ src/main/java/com/sparta/voyageblog/service/PostService.java
@@ -4,20 +4,20 @@ import com.sparta.voyageblog.dto.PostRequestDto;
 import com.sparta.voyageblog.dto.PostResponseDto;
 import com.sparta.voyageblog.dto.PostUpdateRequestDto;
 import com.sparta.voyageblog.entity.Post;
+import com.sparta.voyageblog.entity.User;
 import com.sparta.voyageblog.repository.PostRepository;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Objects;
 
 @Service
 public class PostService {
 
     private final PostRepository postRepository;
 
-    public PostService(PostRepository postRepository){
-        this.postRepository=postRepository;
+    public PostService(PostRepository postRepository) {
+        this.postRepository = postRepository;
     }
 
     public List<PostResponseDto> getPosts() {
@@ -25,49 +25,39 @@ public class PostService {
     }
 
     //Post 생성
-    public PostResponseDto createPost(PostRequestDto requestDto) {
-        Post post= new Post(requestDto);
-        Post savePost= postRepository.save(post);
-        return new PostResponseDto(savePost);
+    public PostResponseDto createPost(PostRequestDto requestDto, User user) {
+        return new PostResponseDto(postRepository.save(new Post(requestDto, user)));
     }
-    
+
     //특정 id의 post 조회
     public PostResponseDto getPostById(Long id) {
-         Post findPost=findPost(id);
-         return new PostResponseDto(findPost);
+        return new PostResponseDto(findPost(id));
     }
 
     //Post 수정
     @Transactional
-    public PostResponseDto updatePost(PostUpdateRequestDto requestDto) {
-        Post post=findPost(requestDto.getId()); //존재찾기
-
-        /*if(checkPassword(post.getPassword(),requestDto.getPassword())){//비밀번호 확인되면
-            post.update(requestDto); //dirty check 로 update 된다.
-        }*/
-        //return new PostResponseDto(findPost(requestDto.getId())); //다시 불러와서 조회
+    public PostResponseDto updatePost(PostUpdateRequestDto requestDto, User user) {
+        Post post = findPost(requestDto.getId()); //기존 것
+        if(!post.getUsername().equals(user.getUsername())){ //기존 게시글의 username과 현재 로그인한 username 비교해서 다르면 예외
+            throw new IllegalArgumentException("회원님이 작성한 게시글이 아닙니다. 수정할 수 없습니다.");
+        }
+        post.updatePost(requestDto);
         return new PostResponseDto(post);
     }
+
     @Transactional
-    public Boolean deletePost(Long id) {
+    public PostResponseDto deletePost(Long id, String username) {
         Post post = findPost(id);
-        /*if (checkPassword(password, post.getPassword())){
-            postRepository.delete(post);
-            return true;
-        }else {
-            return false;
-        }*/
-        return true;
+        if (!post.getUsername().equals(username)) {
+            throw new IllegalArgumentException("회원님이 작성한 게시글이 아닙니다. 삭제할 수 없습니다.");
+        }
+        postRepository.delete(post);
+        return new PostResponseDto(post);
     }
 
     //private methods Service 내에서만 사용되는 메서드!
-    //비밀번호 일치 여부 확인 Boolean
-    private Boolean checkPassword(String input,String password) {
-        return Objects.equals(input, password);
-    }
-
     //특정 id를 가지는 Post 를 찾는 method, getPostById와 다른점은 리턴값이 post 객체라는 점
-    private Post findPost(Long id){
+    private Post findPost(Long id) {
         return postRepository.findById(id).orElseThrow(() ->
                 new IllegalArgumentException("선택한 게시글은 존재하지 않습니다.")
         );
